Develop2

+++ src/components/Detail/Comment/Comment.jsx
@@ -45,7 +45,6 @@ function Comment() {
 	const [comment, setComment] = useState({ comment: content });
 
 	const userInfo = useSelector(state => state.mypage?.userInfo);
-	const companyInfo = useSelector(state => state.mypage?.companyInfo);
 
 	useEffect(() => {
 		dispatch(__getComment(id));
@@ -170,7 +169,11 @@ function Comment() {
 												)}
 												<Div
 													onClick={() => {
-														dispatch(__deleteComment(commentId));
+														if (window.confirm("댓글을 삭제하시겠습니까?")) {
+															dispatch(__deleteComment(commentId));
+														} else {
+															alert("취소되었습니다.");
+														}
 													}}
 												>
 													삭제
